requirements.txt
flask>=3.0.0
pyyaml>=6.0.3
python-dotenv>=1.2.0
requests>=2.32.5
flask-hot-reload>=0.3.0
watchdog>=6.0.0
